main, parsers, refactor: remove "meta-" prefix from header files
I'm renewing terms for combining parsers for an input file.
What I called "meta parser" is now called "base parser".
So "meta-" prefix in header file names doesn't make sense.
