+++ parsers/cxx/cxx_parser_block.c
@@ -18,7 +18,7 @@
 
 #include "parse.h"
 #include "vstring.h"
-#include "../meta-cpreprocessor.h"
+#include "../cpreprocessor.h"
 #include "debug.h"
 #include "keyword.h"
 #include "read.h"
