+++ parsers/cxx/cxx_parser_function.c
@@ -17,7 +17,7 @@
 
 #include "parse.h"
 #include "vstring.h"
-#include "../meta-cpreprocessor.h"
+#include "../cpreprocessor.h"
 #include "debug.h"
 #include "keyword.h"
 #include "read.h"
