+++ parsers/dts.c
@@ -11,7 +11,7 @@
 *   INCLUDE FILES
 */
 #include "general.h"
-#include "meta-cpreprocessor.h"
+#include "cpreprocessor.h"
 #include "kind.h"
 #include "parse.h"
 #include "routines.h"
