+++ parsers/m4.c
@@ -16,7 +16,7 @@
 #include "entry.h"
 #include "htable.h"
 #include "keyword.h"
-#include "meta-m4.h"
+#include "m4.h"
 #include "options.h"
 #include "parse.h"
 #include "read.h"
