+++ parsers/yaml.c
@@ -16,7 +16,7 @@
 #include "parse.h"
 #include "read.h"
 #include "types.h"
-#include "meta-yaml.h"
+#include "yaml.h"
