+++ win32/ctags_vs2013.vcxproj
@@ -244,7 +244,7 @@
     <ClInclude Include="..\main\htable.h" />
     <ClInclude Include="..\main\keyword.h" />
     <ClInclude Include="..\main\kind.h" />
-    <ClInclude Include="..\parsers\meta-cpreprocessor.h" />
+    <ClInclude Include="..\parsers\cpreprocessor.h" />
     <ClInclude Include="..\main\main.h" />
     <ClInclude Include="..\main\nestlevel.h" />
     <ClInclude Include="..\main\options.h" />
@@ -274,9 +274,9 @@
     <ClInclude Include="..\main\promise.h" />
     <ClInclude Include="..\main\dependency.h" />
     <ClInclude Include="..\main\gcc-attr.h" />
-    <ClInclude Include="..\parsers\meta-make.h" />
-    <ClInclude Include="..\parsers\meta-iniconf.h" />
-    <ClInclude Include="..\parsers\meta-m4.h" />
+    <ClInclude Include="..\parsers\make.h" />
+    <ClInclude Include="..\parsers\iniconf.h" />
+    <ClInclude Include="..\parsers\m4.h" />
     <ClInclude Include="..\main\ptrarray.h" />
     <ClInclude Include="..\main\objpool.h" />
     <ClInclude Include="..\main\tokeninfo.h" />
